# contracts/fabric-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OPL Fabric API",
    "version": "0.1.0",
    "license": {
      "name": "Proprietary",
      "url": "https://github.com/RenDeHuang/OPL-Fabric"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787"
    }
  ],
  "security": [
    {
      "operatorToken": []
    }
  ],
  "components": {
    "securitySchemes": {
      "operatorToken": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Readiness": {
        "type": "object",
        "required": ["ready", "provider", "missingEnv", "resourceCatalog", "blockers", "repairHints"],
        "properties": {
          "ready": { "type": "boolean" },
          "provider": { "type": "string", "minLength": 1 },
          "missingEnv": { "type": "array", "items": { "type": "string", "minLength": 1 } },
          "resourceCatalog": { "$ref": "#/components/schemas/Catalog" },
          "blockers": { "type": "array", "items": { "type": "string", "minLength": 1 } },
          "repairHints": { "type": "array", "items": { "type": "string", "minLength": 1 } }
        }
      },
      "Catalog": {
        "type": "object",
        "required": ["schemaVersion", "owner", "productPresets", "computeProfiles", "providerInstanceTypes", "capacityPools", "schedulingPolicies", "storageClasses", "workspaceImages", "ingressDomains"],
        "properties": {
          "schemaVersion": { "type": "integer", "minimum": 2 },
          "owner": { "type": "string", "minLength": 1 },
          "productPresets": { "type": "array", "items": { "$ref": "#/components/schemas/ProductPreset" } },
          "computeProfiles": { "type": "array", "items": { "$ref": "#/components/schemas/ComputeProfile" } },
          "providerInstanceTypes": { "type": "array", "items": { "$ref": "#/components/schemas/ProviderInstanceType" } },
          "capacityPools": { "type": "array", "items": { "$ref": "#/components/schemas/CapacityPool" } },
          "schedulingPolicies": { "type": "array", "items": { "$ref": "#/components/schemas/SchedulingPolicy" } },
          "storageClasses": { "type": "array", "items": { "$ref": "#/components/schemas/StorageClass" } },
          "workspaceImages": { "type": "array", "items": { "$ref": "#/components/schemas/WorkspaceImage" } },
          "ingressDomains": { "type": "array", "items": { "$ref": "#/components/schemas/IngressDomain" } }
        }
      },
      "ProductPreset": {
        "type": "object",
        "required": ["id", "name", "accelerator", "defaultCpu", "defaultMemoryGb", "defaultGpu", "defaultDiskGb", "available", "computeProfileId", "storageClassId", "workspaceImageId", "ingressDomainId", "schedulingPolicyId"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "name": { "type": "string", "minLength": 1 },
          "accelerator": { "type": "string", "enum": ["cpu", "gpu"] },
          "defaultCpu": { "type": "integer", "minimum": 1 },
          "defaultMemoryGb": { "type": "integer", "minimum": 1 },
          "defaultGpu": { "type": "integer", "minimum": 0 },
          "defaultDiskGb": { "type": "integer", "minimum": 1 },
          "available": { "type": "boolean" },
          "unavailableReason": { "type": "string" },
          "computeProfileId": { "type": "string", "minLength": 1 },
          "storageClassId": { "type": "string", "minLength": 1 },
          "workspaceImageId": { "type": "string", "minLength": 1 },
          "ingressDomainId": { "type": "string", "minLength": 1 },
          "schedulingPolicyId": { "type": "string", "minLength": 1 }
        }
      },
      "ComputeProfile": {
        "type": "object",
        "required": ["id", "name", "accelerator", "provider", "minCpu", "maxCpu", "minMemoryGb", "maxMemoryGb", "maxGpu", "customShapesAllowed", "available"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "name": { "type": "string", "minLength": 1 },
          "accelerator": { "type": "string", "enum": ["cpu", "gpu"] },
          "provider": { "type": "string", "minLength": 1 },
          "minCpu": { "type": "integer", "minimum": 1 },
          "maxCpu": { "type": "integer", "minimum": 1 },
          "minMemoryGb": { "type": "integer", "minimum": 1 },
          "maxMemoryGb": { "type": "integer", "minimum": 1 },
          "maxGpu": { "type": "integer", "minimum": 0 },
          "customShapesAllowed": { "type": "boolean" },
          "available": { "type": "boolean" },
          "unavailableReason": { "type": "string" }
        }
      },
      "ProviderInstanceType": {
        "type": "object",
        "required": ["id", "provider", "region", "family", "instanceType", "accelerator", "minCpu", "maxCpu", "minMemoryGb", "maxMemoryGb", "maxGpu", "capacityPoolIds", "available"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "provider": { "type": "string", "minLength": 1 },
          "region": { "type": "string", "minLength": 1 },
          "zone": { "type": "string" },
          "family": { "type": "string", "minLength": 1 },
          "instanceType": { "type": "string", "minLength": 1 },
          "accelerator": { "type": "string", "enum": ["cpu", "gpu"] },
          "minCpu": { "type": "integer", "minimum": 1 },
          "maxCpu": { "type": "integer", "minimum": 1 },
          "minMemoryGb": { "type": "integer", "minimum": 1 },
          "maxMemoryGb": { "type": "integer", "minimum": 1 },
          "maxGpu": { "type": "integer", "minimum": 0 },
          "capacityPoolIds": { "type": "array", "items": { "type": "string", "minLength": 1 } },
          "available": { "type": "boolean" },
          "priceHint": { "type": "string" }
        }
      },
      "CapacityPool": {
        "type": "object",
        "required": ["id", "provider", "kind", "isolationMode", "region", "computeProfileIds", "available"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "provider": { "type": "string", "minLength": 1 },
          "kind": { "type": "string", "enum": ["nodepool_by_instance_type"] },
          "isolationMode": { "type": "string", "enum": ["workspace_exclusive_cvm"] },
          "region": { "type": "string", "minLength": 1 },
          "zone": { "type": "string" },
          "providerRef": { "type": "string" },
          "computeProfileIds": { "type": "array", "items": { "type": "string", "minLength": 1 } },
          "available": { "type": "boolean" },
          "unavailableReason": { "type": "string" }
        }
      },
      "SchedulingPolicy": {
        "type": "object",
        "required": ["id", "mode", "capacityPoolKind", "createPoolIfMissing", "rejectIfNoCapacity", "description"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "mode": { "type": "string", "enum": ["workspace_exclusive_cvm"] },
          "capacityPoolKind": { "type": "string", "enum": ["nodepool_by_instance_type"] },
          "createPoolIfMissing": { "type": "boolean" },
          "rejectIfNoCapacity": { "type": "boolean" },
          "description": { "type": "string", "minLength": 1 }
        }
      },
      "StorageClass": {
        "type": "object",
        "required": ["id", "provider", "storageClassName", "accessMode", "available"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "provider": { "type": "string", "minLength": 1 },
          "storageClassName": { "type": "string", "minLength": 1 },
          "accessMode": { "type": "string", "minLength": 1 },
          "available": { "type": "boolean" }
        }
      },
      "WorkspaceImage": {
        "type": "object",
        "required": ["id", "image", "port", "persistentMounts", "available"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "image": { "type": "string", "minLength": 1 },
          "port": { "type": "integer", "minimum": 1 },
          "persistentMounts": { "type": "array", "items": { "type": "string", "minLength": 1 } },
          "available": { "type": "boolean" }
        }
      },
      "IngressDomain": {
        "type": "object",
        "required": ["id", "host", "pathPattern", "available"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "host": { "type": "string", "minLength": 1 },
          "pathPattern": { "type": "string", "minLength": 1 },
          "available": { "type": "boolean" }
        }
      },
      "OperationReceipt": {
        "type": "object",
        "required": ["operationId", "state", "resourceKind", "resourceId"],
        "properties": {
          "operationId": { "type": "string", "minLength": 1 },
          "state": { "type": "string", "enum": ["accepted", "dry_run", "applying", "verifying", "succeeded", "failed", "blocked", "needs_human_gate"] },
          "resourceKind": { "type": "string", "minLength": 1 },
          "resourceId": { "type": "string", "minLength": 1 }
        }
      },
      "StorageVolume": {
        "type": "object",
        "required": ["id", "state", "providerRef", "sizeGb", "retained"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "state": { "type": "string", "minLength": 1 },
          "providerRef": { "type": "string" },
          "sizeGb": { "type": "integer", "minimum": 1 },
          "retained": { "type": "boolean" }
        }
      },
      "ComputeAllocation": {
        "type": "object",
        "required": ["id", "state", "providerRef", "runtimeRef"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "state": { "type": "string", "minLength": 1 },
          "providerRef": { "type": "string" },
          "runtimeRef": { "type": "string" }
        }
      },
      "StorageAttachment": {
        "type": "object",
        "required": ["id", "computeAllocationId", "storageId", "state", "mountPath", "providerRef"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "computeAllocationId": { "type": "string", "minLength": 1 },
          "storageId": { "type": "string", "minLength": 1 },
          "state": { "type": "string", "minLength": 1 },
          "mountPath": { "type": "string", "minLength": 1 },
          "providerRef": { "type": "string" }
        }
      },
      "WorkspaceEntry": {
        "type": "object",
        "required": ["id", "workspaceId", "attachmentId", "state", "url", "serviceRef"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "workspaceId": { "type": "string", "minLength": 1 },
          "attachmentId": { "type": "string", "minLength": 1 },
          "state": { "type": "string", "minLength": 1 },
          "url": { "type": "string", "minLength": 1 },
          "serviceRef": { "type": "string" }
        }
      },
      "CreateStorageVolumeRequest": {
        "type": "object",
        "required": ["accountId", "requestedBy", "sizeGb"],
        "properties": {
          "accountId": { "type": "string", "minLength": 1 },
          "requestedBy": { "type": "string", "minLength": 1 },
          "productPresetId": { "type": "string" },
          "sizeGb": { "type": "integer", "minimum": 1 }
        }
      },
      "CreateComputeAllocationRequest": {
        "type": "object",
        "required": ["accountId", "requestedBy"],
        "properties": {
          "accountId": { "type": "string", "minLength": 1 },
          "requestedBy": { "type": "string", "minLength": 1 },
          "productPresetId": { "type": "string" },
          "computeShape": { "type": "object", "additionalProperties": true },
          "providerInstanceType": { "type": "string" },
          "capacityPoolId": { "type": "string" },
          "isolationMode": { "type": "string", "enum": ["", "workspace_exclusive_cvm"] }
        }
      },
      "CreateStorageAttachmentRequest": {
        "type": "object",
        "required": ["accountId", "requestedBy", "computeAllocationId", "storageId"],
        "properties": {
          "accountId": { "type": "string", "minLength": 1 },
          "requestedBy": { "type": "string", "minLength": 1 },
          "computeAllocationId": { "type": "string", "minLength": 1 },
          "storageId": { "type": "string", "minLength": 1 },
          "mountPath": { "type": "string" }
        }
      },
      "CreateWorkspaceEntryRequest": {
        "type": "object",
        "required": ["accountId", "requestedBy", "workspaceName", "attachmentId"],
        "properties": {
          "accountId": { "type": "string", "minLength": 1 },
          "requestedBy": { "type": "string", "minLength": 1 },
          "workspaceName": { "type": "string", "minLength": 1 },
          "attachmentId": { "type": "string", "minLength": 1 }
        }
      },
      "CreateWorkspaceRequest": {
        "type": "object",
        "required": ["accountId", "requestedBy", "workspaceName"],
        "properties": {
          "accountId": { "type": "string", "minLength": 1 },
          "requestedBy": { "type": "string", "minLength": 1 },
          "workspaceName": { "type": "string", "minLength": 1 },
          "productPresetId": { "type": "string" },
          "computeShape": { "type": "object", "additionalProperties": true },
          "providerInstanceType": { "type": "string" },
          "capacityPoolId": { "type": "string" },
          "isolationMode": { "type": "string", "enum": ["", "workspace_exclusive_cvm"] },
          "storage": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "sizeGb": { "type": "integer", "minimum": 1 }
            }
          }
        }
      },
      "Workspace": {
        "type": "object",
        "required": ["workspaceId", "state", "storage", "computeAllocation", "attachment", "entry", "operationId"],
        "properties": {
          "workspaceId": { "type": "string", "minLength": 1 },
          "state": { "type": "string", "enum": ["provisioning", "running", "blocked", "failed", "destroyed"] },
          "operationId": { "type": "string", "minLength": 1 },
          "storage": {
            "type": "object",
            "required": ["id", "state", "sizeGb"],
            "properties": {
              "id": { "type": "string", "minLength": 1 },
              "state": { "type": "string", "minLength": 1 },
              "sizeGb": { "type": "integer", "minimum": 1 }
            }
          },
          "computeAllocation": {
            "type": "object",
            "required": ["id", "state"],
            "properties": {
              "id": { "type": "string", "minLength": 1 },
              "state": { "type": "string", "minLength": 1 }
            }
          },
          "attachment": {
            "type": "object",
            "required": ["id", "state", "mountPath"],
            "properties": {
              "id": { "type": "string", "minLength": 1 },
              "state": { "type": "string", "minLength": 1 },
              "mountPath": { "type": "string", "minLength": 1 }
            }
          },
          "entry": {
            "type": "object",
            "required": ["id", "state", "url"],
            "properties": {
              "id": { "type": "string", "minLength": 1 },
              "state": { "type": "string", "minLength": 1 },
              "url": { "type": "string", "minLength": 1 }
            }
          }
        }
      },
      "ConfirmRequest": {
        "type": "object",
        "required": ["requestedBy", "confirm"],
        "properties": {
          "requestedBy": { "type": "string", "minLength": 1 },
          "confirm": { "type": "boolean", "const": true }
        }
      }
    }
  },
  "paths": {
    "/api/fabric/readiness": {
      "get": {
        "operationId": "getFabricReadiness",
        "summary": "Get Fabric readiness",
        "responses": {
          "200": {
            "description": "Fabric readiness",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Readiness" }
              }
            }
          },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/fabric/catalog": {
      "get": {
        "operationId": "getFabricCatalog",
        "summary": "Get Fabric resource catalog",
        "responses": {
          "200": {
            "description": "Fabric resource catalog",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Catalog" }
              }
            }
          },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/fabric/storage-volumes": {
      "post": {
        "operationId": "createFabricStorageVolume",
        "summary": "Accept storage volume creation",
        "parameters": [
          { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "X-Correlation-Id", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateStorageVolumeRequest" } } } },
        "responses": {
          "202": { "description": "Accepted operation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OperationReceipt" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/storage-volumes/{id}": {
      "get": {
        "operationId": "getFabricStorageVolume",
        "summary": "Get Fabric storage volume",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "responses": {
          "200": { "description": "Fabric storage volume", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StorageVolume" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/compute-allocations": {
      "post": {
        "operationId": "createFabricComputeAllocation",
        "summary": "Accept compute allocation creation",
        "parameters": [
          { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "X-Correlation-Id", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateComputeAllocationRequest" } } } },
        "responses": {
          "202": { "description": "Accepted operation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OperationReceipt" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/compute-allocations/{id}": {
      "get": {
        "operationId": "getFabricComputeAllocation",
        "summary": "Get Fabric compute allocation",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "responses": {
          "200": { "description": "Fabric compute allocation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ComputeAllocation" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/storage-attachments": {
      "post": {
        "operationId": "createFabricStorageAttachment",
        "summary": "Accept storage attachment creation",
        "parameters": [
          { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "X-Correlation-Id", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateStorageAttachmentRequest" } } } },
        "responses": {
          "202": { "description": "Accepted operation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OperationReceipt" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/storage-attachments/{id}": {
      "get": {
        "operationId": "getFabricStorageAttachment",
        "summary": "Get Fabric storage attachment",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "responses": {
          "200": { "description": "Fabric storage attachment", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StorageAttachment" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/workspace-entries": {
      "post": {
        "operationId": "createFabricWorkspaceEntry",
        "summary": "Accept workspace entry creation",
        "parameters": [
          { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "X-Correlation-Id", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateWorkspaceEntryRequest" } } } },
        "responses": {
          "202": { "description": "Accepted operation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OperationReceipt" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/workspace-entries/{id}": {
      "get": {
        "operationId": "getFabricWorkspaceEntry",
        "summary": "Get Fabric workspace entry",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "responses": {
          "200": { "description": "Fabric workspace entry", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WorkspaceEntry" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/workspaces": {
      "post": {
        "operationId": "createFabricWorkspace",
        "summary": "Accept workspace delivery",
        "parameters": [
          { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "X-Correlation-Id", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateWorkspaceRequest" } } } },
        "responses": {
          "202": { "description": "Accepted operation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OperationReceipt" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/workspaces/{id}": {
      "get": {
        "operationId": "getFabricWorkspace",
        "summary": "Get Fabric workspace",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "responses": {
          "200": { "description": "Fabric workspace", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Workspace" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/compute-allocations/{id}/destroy": {
      "post": {
        "operationId": "destroyFabricComputeAllocation",
        "summary": "Accept compute destruction",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "X-Correlation-Id", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfirmRequest" } } } },
        "responses": {
          "202": { "description": "Accepted operation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OperationReceipt" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/storage-volumes/{id}/destroy": {
      "post": {
        "operationId": "destroyFabricStorageVolume",
        "summary": "Accept storage destruction",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "X-Correlation-Id", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfirmRequest" } } } },
        "responses": {
          "202": { "description": "Accepted operation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OperationReceipt" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/storage-attachments/{id}/detach": {
      "post": {
        "operationId": "detachFabricStorageAttachment",
        "summary": "Accept storage attachment detach",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } },
          { "name": "X-Correlation-Id", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfirmRequest" } } } },
        "responses": {
          "202": { "description": "Accepted operation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OperationReceipt" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    },
    "/api/fabric/operations/{id}": {
      "get": {
        "operationId": "getFabricOperation",
        "summary": "Get Fabric operation",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "responses": {
          "200": { "description": "Fabric operation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OperationReceipt" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "503": { "description": "Store unavailable" }
        }
      }
    }
  }
}